requirements.txt
async-timeout==4.0.3
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
EasyProcess==1.1
entrypoint2==1.1
flask==2.3.3
Flask-Cors==4.0.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
MouseInfo==0.1.3
mss==9.0.1
Pillow==10.0.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dotenv==1.0.0
pytweening==1.0.7
redis==5.0.1
requests==2.31.0
urllib3==2.0.5
werkzeug==2.3.7
zipp==3.17.0
